+++ src/datarefinery/plugins/tabular/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2026 Pointmatic
+# SPDX-License-Identifier: Apache-2.0
+
+from datarefinery.plugins.tabular.plugin import PLUGIN, TabularPlugin
+
+__all__ = ["PLUGIN", "TabularPlugin"]