+++ src/datarefinery/plugins/text/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2026 Pointmatic
+# SPDX-License-Identifier: Apache-2.0
+
+from datarefinery.plugins.text.plugin import PLUGIN, TextPlugin
+
+__all__ = ["PLUGIN", "TextPlugin"]